refactor: 멤버 및 조직, 영상 관련 검증 메서드 이동
