+++ src/main/java/app/allstackproject/privideo/repository/comment/CommentRepositoryCustom.java
@@ -5,6 +5,4 @@ import java.util.List;
 
 public interface CommentRepositoryCustom {
     List<CommentInfo> findAllByVideoId(Long videoId);
-
-    boolean isValidMemberAndOrgAndVideo(Long memberId, Long orgId, Long videoId);
 }
