+++ src/main/java/app/allstackproject/privideo/repository/video/custom/VideoRepositoryCustom.java
@@ -1,0 +1,5 @@
+package app.allstackproject.privideo.repository.video.custom;
+
+public interface VideoRepositoryCustom {
+    boolean isValidMemberAndOrgAndVideo(Long memberId, Long orgId, Long videoId);
+}
